requirements.txt
pip>=24.0
setuptools>=70.0
wheel>=0.41.0

fastapi==0.109.0
uvicorn==0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0

numpy>=2.0.0
pandas>=2.1.0
scikit-learn>=1.4.0

sentence-transformers>=3.0.0
transformers>=4.40.0
torch>=2.2.0
spacy>=3.7.0

pdfplumber>=0.10.0
python-docx>=1.0.0
nltk>=3.8.0

streamlit>=1.32.0
python-dotenv>=1.0.0
requests>=2.31.0